feat(rust): implement `FunctionExpr` for bound and round methods
