+++ polars/polars-lazy/polars-plan/src/dsl/function_expr/round.rs
@@ -1,0 +1,13 @@
+use super::*;
+
+pub(super) fn round(s: &Series, decimals: u32) -> PolarsResult<Series> {
+    s.round(decimals)
+}
+
+pub(super) fn floor(s: &Series) -> PolarsResult<Series> {
+    s.floor()
+}
+
+pub(super) fn ceil(s: &Series) -> PolarsResult<Series> {
+    s.ceil()
+}
