+++ polars/polars-lazy/polars-plan/src/dsl/function_expr/schema.rs
@@ -296,6 +296,9 @@ impl FunctionExpr {
             #[cfg(feature = "log")]
             Entropy { .. } | Log { .. } | Log1p | Exp => float_dtype(),
             Unique(_) => same_type(),
+            #[cfg(feature = "round_series")]
+            Round { .. } | Floor | Ceil => same_type(),
+            UpperBound | LowerBound => same_type(),
         }
     }
 }
